+++ _pages/profile/[id]/submission-details-accordion.js
@@ -27,6 +27,7 @@ const submissionDetailsAccordionFragments = {
         requester
         arbitrator
         arbitratorExtraData
+        id
         evidence(orderBy: creationTime, orderDirection: desc) {
           creationTime
           id
@@ -38,6 +39,7 @@ const submissionDetailsAccordionFragments = {
           reason
           disputeID
           challenger
+          challengeID
           rounds(
             orderBy: creationTime
             orderDirection: asc
@@ -87,6 +89,7 @@ export default function SubmissionDetailsAccordion({ submission, contract }) {
     loserStakeMultiplier,
   } = useFragment(submissionDetailsAccordionFragments.contract, contract);
   const { web3 } = useWeb3();
+
   return (
     <Accordion>
       <SubmissionDetailsAccordionItem